A lot of random changes ... sorry future me
I keep going back and forth on how to handle errors and what types I
should be using. I think I'm going to stick with what I have now ...
hopefully.

Also started to figure out how to build a graph (trie) so that I can
have a better way for handling key presses. I think the implementation
is good enough for now but it's not complete nor fully tested. Just
wanted to commit cause I kept wanting to make more changes ... honestly
should have broke this up into different commits but I am being
extremely lazy today *<:o)

+++ src/screen.h
@@ -10,14 +10,14 @@ typedef struct Padding {
 } padding_t;
 
 typedef struct Cursor {
-    u_int32_t row;
-    u_int32_t col;
+    size_t row;
+    size_t col;
 } cursor_t;
 
 typedef struct Screen {
     cursor_t cursor;
-    u_int32_t max_row;
-    u_int32_t max_col;
+    size_t max_row;
+    size_t max_col;
     padding_t padding;
 } screen_t;
 
@@ -30,14 +30,14 @@ typedef enum DIRECTION {
     LEFT = 'D'
 } direction_e;
 
-u_int32_t screen_get_row_index(screen_t *screen);
-u_int32_t screen_get_row_count(screen_t *screen);
-u_int32_t screen_get_row_max(screen_t *screen);
-u_int32_t screen_get_row_min(screen_t *screen);
+size_t screen_get_row_index(screen_t *screen);
+size_t screen_get_row_count(screen_t *screen);
+size_t screen_get_row_max(screen_t *screen);
+size_t screen_get_row_min(screen_t *screen);
 
-u_int32_t screen_get_column_index(screen_t *screen);
-u_int32_t screen_get_column_count(screen_t *screen);
-u_int32_t screen_get_column_max(screen_t *screen);
-u_int32_t screen_get_column_min(screen_t *screen);
+size_t screen_get_column_index(screen_t *screen);
+size_t screen_get_column_count(screen_t *screen);
+size_t screen_get_column_max(screen_t *screen);
+size_t screen_get_column_min(screen_t *screen);
 
 void screen_move_cursor(screen_t *screen, direction_e direction);
